+++ src/modules/screens/screen_diag.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "display_ble_context.h"
+class V1Display;
+void drawDiagScreen(V1Display& disp, const DisplayBleContext& ble);