+++ src/modules/wifi/wifi_jbv1_api_service.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <WebServer.h>
+
+namespace WifiJBV1ApiService {
+    void handleApiUpdate(WebServer& server);
+}